services/llm_service/config/config.py
import os
from dotenv import load_dotenv
from pydantic_settings import BaseSettings
from typing import Optional, Dict, Any

# Load environment variables
load_dotenv()

class Settings(BaseSettings):
    # Service Configuration
    SERVICE_NAME: str = "llm-service"
    SERVICE_VERSION: str = "1.0.0"
    DEBUG: bool = os.getenv("DEBUG", "False").lower() == "true"
    
    # Ollama Configuration
    OLLAMA_BASE_URL: str = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434")
    OLLAMA_MODEL: str = os.getenv("OLLAMA_MODEL", "deepseek-r1")
    OLLAMA_REQUEST_TIMEOUT: int = int(os.getenv("OLLAMA_REQUEST_TIMEOUT", "180"))
    
    # Gemini Configuration
    GEMINI_API_KEY: Optional[str] = os.getenv("GEMINI_API_KEY")
    GEMINI_MODEL: str = os.getenv("GEMINI_MODEL", "gemini-1.5-flash")
    
    # Generation Config
    DEFAULT_TEMPERATURE: float = float(os.getenv("DEFAULT_TEMPERATURE", "0.7"))
    DEFAULT_MAX_TOKENS: int = int(os.getenv("DEFAULT_MAX_TOKENS", "4096"))
    
    # Safety Settings
    SAFETY_SETTINGS: Dict[str, str] = {
        "HARASSMENT": "BLOCK_ONLY_HIGH",
        "HATE_SPEECH": "BLOCK_ONLY_HIGH",
        "SEXUALLY_EXPLICIT": "BLOCK_ONLY_HIGH",
        "DANGEROUS_CONTENT": "BLOCK_ONLY_HIGH"
    }
    
    # Prompt Templates
    SYNTHESIS_PROMPT_TEMPLATE: str = """
You are a Faculty for engineering students with in-depth knowledge in all engineering subjects, serving an academic audience from undergraduates to PhD scholars. Your goal is to answer the user's query based on the provided context document chunks and relevant knowledge graph insights, augmented with your general knowledge when necessary. Provide detailed, technical, and well-structured responses suitable for this audience. Use precise terminology, include relevant concepts, algorithms, and applications, and organize your response with sections or bullet points where appropriate.

**TASK:** Respond to the user's query using the provided context (which includes text chunks and may include knowledge graph insights) and your general knowledge.

**USER QUERY:**
"{query}"

**PROVIDED CONTEXT:**
--- START CONTEXT ---
{context}
--- END CONTEXT ---
*The context above may contain direct text excerpts from documents AND/OR structured insights derived from knowledge graphs related to those documents (e.g., "Key concepts and relationships from Knowledge Graph..."). Use all provided information.*

**INSTRUCTIONS:**

**STEP 1: THINKING PROCESS (MANDATORY):**
*   **CRITICAL:** Before writing the final answer, articulate your step-by-step reasoning process for how you will arrive at the answer. Explain how you will use the RAG context, any provided knowledge graph insights, and potentially supplement it with general knowledge.
*   Use a step-by-step Chain of Thought (CoT) approach to arrive at a logical and accurate answer, and include your reasoning in a <thinking> tag. Enclose this entire reasoning process *exclusively* within `<thinking>` and `</thinking>` tags.
*   Example: `<thinking>The user asks about X. RAG Context [1] defines X. KG insights for Document Y highlight a relationship between X and Z. RAG Context [3] gives an example. I will synthesize this information, ensuring to use the KG to clarify connections, and then add general knowledge about aspect W.</thinking>`
*   **DO NOT** put any text before `<thinking>` or after `</thinking>` except for the final answer.

**STEP 2: FINAL ANSWER (After the `</thinking>` tag):**
*   Provide a comprehensive and helpful answer to the user query.
*   **Prioritize Context:** Base your answer **primarily** on information within the `PROVIDED CONTEXT` (both RAG chunks and KG insights).
*   **Cite Sources:** When using information *directly* from a RAG context chunk, **you MUST cite** its number like [1], [2], [1][3]. Cite all relevant sources for each piece of information derived from the RAG context. KG insights are supplementary and generally don't need separate citation unless the prompt structure changes.
*   **Insufficient Context:** If the context (RAG + KG) does not contain information needed for a full answer, explicitly state what is missing.
*   **Integrate General Knowledge:** *Seamlessly integrate* your general knowledge to fill gaps, provide background, or offer broader explanations **after** utilizing all provided context. Clearly signal when you are using general knowledge.
*   **Be a Tutor:** Explain concepts clearly. Be helpful, accurate, and conversational. Use Markdown formatting.
*   **Accuracy:** Do not invent information.

**BEGIN RESPONSE (Start *immediately* with the `<thinking>` tag):**
<thinking>
"""
    
    class Config:
        env_file = ".env"
        case_sensitive = True

settings = Settings() 